# openapi.yaml
openapi: 3.0.0
info:
  title: Pathguard API
  version: 1.0.0
paths:
  /api/v1/health:
    get:
      tags:
        - API Health
      summary: Get API health status
      description: Retrieve the health status and uptime information of the PathGuard API.
      operationId: getApiHealth
      responses:
        "200":
          description: Successful response with API health status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  timestamp:
                    type: integer
                    example: 1759239042808
  /api/v1:
    get:
      tags:
        - API Info
      summary: Get API information
      description: >-
        Retrieve basic information about the PathGuard API, including available
        endpoints and version.
      operationId: getApiInfo
      responses:
        "200":
          description: Successful response with API information
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Welcome to PathGuard API.
                  version:
                    type: string
                    example: ^1.0.0
                  endpoints:
                    type: object
                    properties:
                      health:
                        type: string
                        example: /api/v1/health
                      kyc:
                        type: string
                        example: /api/v1/kyc
                      docs:
                        type: string
                        example: /api/v1/docs
  /api/v1/kyc:
    get:
      tags:
        - KYC
      summary: Get KYC endpoint information
      description: >-
        Retrieve a list of all KYC verification endpoints, available methods,
        and documentation links.
      responses:
        "200":
          description: Successful response with KYC endpoint information
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: KYC service endpoints
                  endpoints:
                    type: array
                    items:
                      type: object
                      properties:
                        path:
                          type: string
                          example: /api/v1/kyc/verifications
                        method:
                          type: string
                          example: POST
                        description:
                          type: string
                          example: Perform identity verification
                    example:
                      - path: /api/v1/kyc/verifications
                        method: POST
                        description: Perform identity verification
                      - path: /api/v1/kyc/verifications
                        method: GET
                        description: List all verifications
                      - path: /api/v1/kyc/verifications/{id}
                        method: GET
                        description: Get verification by ID
                  documentation:
                    type: string
                    example: /api/v1/docs/#tag/kyc
  /api/v1/kyc/verifications:
    post:
      tags:
        - KYC
      summary: Perform identity verification
      description: Perform KYC verification with risk scoring and validation checks.
      operationId: verifyKyc
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - documentType
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 100
                  description: Full name of the person (2-100 characters)
                  example: Alice Dupont
                email:
                  type: string
                  format: email
                  description: Valid email address (will be converted to lowercase)
                  example: alice.dupont@company.com
                documentType:
                  type: string
                  enum:
                    - passport
                    - id_card
                    - driver_license
                  description: Type of identification document
                  example: id_card
      responses:
        "200":
          description: Verification completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  verification_id:
                    type: string
                    pattern: >-
                      ^ver_[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$
                    description: Unique verification identifier with 'ver_' prefix
                    example: ver_123e4567-e89b-12d3-a456-426614174000
                  status:
                    type: string
                    enum:
                      - approved
                      - pending_review
                      - rejected
                    description: >-
                      Verification status: approved (score ≤ 0.2),
                      pending_review (0.2 < score ≤ 0.5), rejected (score > 0.5)
                    example: approved
                  risk_score:
                    type: number
                    format: float
                    minimum: 0
                    maximum: 1
                    description: >-
                      Calculated risk score based on name length, email
                      validity, and document type
                    example: 0.1
                  checks_performed:
                    type: array
                    items:
                      type: string
                    description: List of validation checks performed during verification
                    example:
                      - name
                      - email
                      - documentType
                  timestamp:
                    type: string
                    format: date-time
                    description: ISO 8601 timestamp of when the verification was processed
                    example: "2025-09-30T10:30:00.000Z"
        "400":
          description: Validation error - invalid request data
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Validation failed
                  details:
                    type: array
                    items:
                      type: object
                      properties:
                        field:
                          type: string
                          description: The field that failed validation
                          example: name
                        message:
                          type: string
                          description: Detailed validation error message
                          example: Name must be at least 2 characters long
                    example:
                      - field: name
                        message: Name must be at least 2 characters long
                      - field: email
                        message: Invalid email format
                      - field: documentType
                        message: >-
                          Invalid option: expected one of
                          "passport"|"id_card"|"driver_license"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Internal server error
    get:
      tags:
        - KYC
      summary: List all verifications
      description: >-
        Retrieve a list of all KYC verifications with optional pagination
        support.
      operationId: getVerifications
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
          required: false
          description: Maximum number of verifications to return
          example: 50
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
          required: false
          description: Number of verifications to skip for pagination
          example: 0
      responses:
        "200":
          description: List of verifications retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    verification_id:
                      type: string
                      pattern: >-
                        ^ver_[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$
                      description: Unique verification identifier with 'ver_' prefix
                      example: ver_123e4567-e89b-12d3-a456-426614174000
                    status:
                      type: string
                      enum:
                        - approved
                        - pending_review
                        - rejected
                      description: >-
                        Verification status: approved (score ≤ 0.2),
                        pending_review (0.2 < score ≤ 0.5), rejected (score >
                        0.5)
                      example: approved
                    risk_score:
                      type: number
                      format: float
                      minimum: 0
                      maximum: 1
                      description: >-
                        Calculated risk score based on name length, email
                        validity, and document type
                      example: 0.1
                    created_at:
                      type: string
                      format: date-time
                      description: ISO 8601 timestamp of when the verification was created
                      example: "2025-10-05T14:30:00.000Z"
              example:
                - verification_id: ver_123e4567-e89b-12d3-a456-426614174000
                  status: approved
                  risk_score: 0.1
                  created_at: "2025-10-05T14:30:00.000Z"
                - verification_id: ver_987fcdeb-51a2-43f1-b987-123456789abc
                  status: pending_review
                  risk_score: 0.35
                  created_at: "2025-10-05T12:45:00.000Z"
                - verification_id: ver_456def78-90ab-12cd-34ef-567890abcdef
                  status: rejected
                  risk_score: 0.6
                  created_at: "2025-10-05T11:20:00.000Z"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Internal server error
  /api/v1/kyc/verifications/{id}:
    get:
      tags:
        - KYC
      summary: Get verification by ID
      description: >-
        Retrieve detailed information about a KYC verification using its unique
        identifier.
      operationId: getVerificationById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Unique verification identifier with 'ver_' prefix followed by UUID
          example: ver_456def78-90ab-12cd-34ef-567890abcdef
      responses:
        "200":
          description: Verification retrieved successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - verification_id
                  - status
                  - risk_score
                  - created_at
                properties:
                  verification_id:
                    type: string
                    pattern: >-
                      ^ver_[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$
                    description: Unique verification identifier
                    example: ver_456def78-90ab-12cd-34ef-567890abcdef
                  status:
                    type: string
                    enum:
                      - approved
                      - pending_review
                      - rejected
                    description: Verification status based on risk score thresholds
                    example: pending_review
                  risk_score:
                    type: number
                    format: float
                    minimum: 0
                    maximum: 1
                    description: Calculated risk score (0 = lowest risk, 1 = highest risk)
                    example: 0.35
                  created_at:
                    type: string
                    format: date-time
                    description: Timestamp when the verification was created
                    example: "2025-10-05T14:30:00.000Z"
              example:
                verification_id: ver_456def78-90ab-12cd-34ef-567890abcdef
                status: pending_review
                risk_score: 0.35
                created_at: "2025-10-05T12:45:00.000Z"
        "400":
          description: Invalid verification ID format
          content:
            application/json:
              schema:
                type: object
                required:
                  - error
                  - details
                properties:
                  error:
                    type: string
                    example: Validation failed
                  details:
                    type: array
                    items:
                      type: object
                      required:
                        - field
                        - message
                      properties:
                        field:
                          type: string
                          example: id
                        message:
                          type: string
                          example: >-
                            Invalid verification ID format. Must be in format
                            ver_xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx
              example:
                error: Validation failed
                details:
                  - field: id
                    message: >-
                      Invalid verification ID format. Must be in format
                      ver_xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx
        "404":
          description: Verification not found
          content:
            application/json:
              schema:
                type: object
                required:
                  - error
                properties:
                  error:
                    type: string
                    example: Verification not found
              example:
                error: Verification not found
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                required:
                  - error
                properties:
                  error:
                    type: string
                    example: Internal server error
              example:
                error: Internal server error
components: {}
tags: []